Replace Streamlit UI with React and FastAPI
