+++ frontend/README.md
@@ -1,0 +1,5 @@
+# React Front-End
+
+This folder contains a minimal React application that can be used to build a user interface for the LLM classification evaluator. The app was created manually and includes only basic files.
+
+Run `npm install` and `npm start` inside this folder to launch the development server. The React app should communicate with the FastAPI backend defined in `backend/api.py`.
