+++ frontend/src/App.js
@@ -1,0 +1,12 @@
+import React from 'react';
+
+function App() {
+  return (
+    <div>
+      <h1>LLM Classification Evaluator</h1>
+      <p>React front-end placeholder</p>
+    </div>
+  );
+}
+
+export default App;
