+++ netlify.toml
@@ -1,15 +1,11 @@
 [build]
-  command = "pip install -r requirements.txt"
-  publish = "."
+  command = "npm run build"
+  publish = "frontend/build"
 
 [build.environment]
-  PYTHON_VERSION = "3.11"
+  NODE_VERSION = "18"
 
 [[redirects]]
   from = "/*"
-  to = "/streamlit_llm_evaluator.py"
+  to = "/index.html"
   status = 200
-
-[context.production.environment]
-  STREAMLIT_SERVER_PORT = "8501"
-  STREAMLIT_SERVER_ADDRESS = "0.0.0.0"
