+++ requirements.txt
@@ -1,8 +1,7 @@
-streamlit>=1.28.0
+fastapi>=0.110.0
+uvicorn>=0.20.0
 pandas>=2.0.0
 numpy>=1.24.0
 openai>=1.0.0
 scikit-learn>=1.3.0
-plotly>=5.15.0
-aiohttp>=3.8.0
 openpyxl>=3.1.0
